# docs/swagger.yaml
basePath: /
definitions:
  handler.AccountView:
    properties:
      channel_id:
        type: string
      concurrent_limit:
        type: integer
      created_at:
        type: string
      current_jobs:
        type: integer
      error_count:
        type: integer
      guild_id:
        type: string
      id:
        type: integer
      is_disabled:
        type: boolean
      is_healthy:
        type: boolean
      last_error:
        type: string
      last_used_at:
        type: string
      success_count:
        type: integer
      updated_at:
        type: string
    type: object
  handler.CreateAccountReq:
    properties:
      channel_id:
        type: string
      guild_id:
        type: string
      user_token:
        type: string
    required:
    - channel_id
    - guild_id
    - user_token
    type: object
...
  handler.UpdateAccountReq:
    properties:
      channel_id:
        type: string
      concurrent_limit:
        type: integer
      guild_id:
        type: string
      is_disabled:
        type: boolean
      user_token:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: string
      data: {}
      detail:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Midjourney Proxy API Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Midjourney Proxy API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      description: Get all account list
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.AccountView'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List accounts
      tags:
      - Account
    post:
      consumes:
      - application/json
      description: Create a new Discord account with the default concurrent limit
        and start its listener. Only guild_id, channel_id, and user_token are accepted.
      parameters:
      - description: Account info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.AccountView'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create account
      tags:
      - Account
  /api/v1/accounts/{id}:
    delete:
      description: Delete the specified account and stop its listener. Accounts with
        active tasks cannot be deleted until current_jobs is 0.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete account
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update the specified Discord account. At least one field must be
        provided. is_healthy is read-only and managed by listener checks; use the
        restart endpoint to re-check an account. While current_jobs is greater than
        0, listener-changing fields (guild_id, channel_id, user_token, and disabling
        the account) are rejected; concurrent_limit can still be changed.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.AccountView'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update account
      tags:
      - Account
  /api/v1/accounts/{id}/restart:
    post:
      description: Restart the specified Discord account listener. is_healthy is managed
        by the server and will be updated after the listener verifies the account.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.AccountView'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Restart account listener
      tags:
      - Account
...
schemes:
- http
- https
swagger: "2.0"